build: Use integration test against the package container
Test the latest build of docker image, instead of binary.

In order to run it as container, `launch-orc` script is introduced to do
bindmount for longhorn-orc, in order to get access to /dev/longhorn/, for
detecting volume purpose.

The IP of longhorn-orc server is passed into integration test through
environment variable now.